% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/key.R
\name{analyze_key}
\alias{analyze_key}
\title{Analyze a key with respect to some data}
\usage{
analyze_key(key, data)
}
\arguments{
\item{key}{key}

\item{data}{data}
}
\value{
List with elements:
  * `common`: character vector, the common part of the key and the data.
  * `rest`: the rest of the *key*.
}
\description{
Analyze a key with respect to some data
}
\keyword{internal}
