% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{cf_set}
\alias{cf_set}
\title{TODO: how does this work?}
\usage{
cf_set(self, private, key, value)
}
\arguments{
\item{self}{Self.}

\item{private}{Private self.}

\item{key}{Key to set.}

\item{value}{Value to set.}
}
\description{
TODO: how does this work?
}
\keyword{internal}
