% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{conf}
\alias{conf}
\alias{conf-package}
\title{Persistent Package Configuration}
\description{
Store the configuration of your package in the user's platform dependent
config file directory. The configuration persists across R sessions,
and can also be edited manually. Configuration files are YAML files.
}
